refactor: string 변환 로직 이동
